Enable mousebites in frame.
Put vertical mousebites in frame to allow it to break away.

A little hackish, and certainly not perfect.  For one, the mousebite
drills seem to pop through the frame too far.  Also, #15 seems to
indicate the better way to do this is to make the frame out of four
rectangles connected together with tabs, this is still full-width frame
all the way around.